Make the capture screen transparent
I'm not sure if this was a cross-platform issue, but it happened anyway
on OS X.

When capturing a screenshot after clicking on 'Capture Area', a gray
screen appeared. Next, when dragging the mouse to select an area, it
didn't show the actual screen area, but a black area.

The cause of this issue was that the JDialog underlying the MyScreen
object was not opaque.