Fix popup messages being cut off
If the message in a Label is too long to fit in its panel, it will get
cut off with three dots. There was some old code in place that tried to
prevent this by calculating the length from the message length, but this
did not work in all cases because it is dependent on many things
including the font.

Using the pack() function instead ensures that the popup's size adapts
to its components' sizes.